requirements.txt
Flask==2.2.2
flask_sqlalchemy==3.0.2
python-dotenv==0.21.0
mariadb==1.1.5.post3
gunicorn==20.1.0